Add files via upload
Modified prevent_starvation method to have class check every PCB if it has been used before. If not, then it will boost priority and make the PCB's flag true notifying the program that its priority has been boosted. As time pass, if the flag is true, it will increment PCB's quantumChk. 

Another method has been created for this class called bring_back_priority where the class will check each PCB in the queue as long as readyQueue is not empty. If quantumChk has reached a certain amount, PCB's priority level goes back to normal. When that happens quantumChk goes back to 0 and flag is false.